refactor: state, action type 정의 및 파일 분리
