+++ client/src/index.tsx
@@ -2,6 +2,10 @@ import React from 'react';
 import ReactDOM from 'react-dom';
 import './index.css';
 import App from './App';
+import { createStore } from 'redux'
+import { rootReducer } from './store';
+
+export const store = createStore(rootReducer)
 
 ReactDOM.render(
   <React.StrictMode>
