+++ client/src/store/index.ts
@@ -1,0 +1,8 @@
+import { combineReducers } from 'redux'
+import { modalReducer } from './modal/reducer'
+
+export const rootReducer = combineReducers({
+  modal: modalReducer
+})
+
+export type RootState = ReturnType<typeof rootReducer>
