+++ client/src/store/modal/actions.ts
@@ -1,0 +1,9 @@
+import { toggleModalState, TOGGLE_MODAL } from "./types"
+
+// action creator
+export const toggleModal = (payload: toggleModalState) => {
+  return {
+    type: TOGGLE_MODAL,
+    payload
+  }
+}
